added payment gateway, customer review, cart operations, wishlist

+++ ecommerce/src/main/java/com/ecom/entity/CartItem.java
@@ -30,9 +30,9 @@ public class CartItem {
 
     private int quantity = 1;
 
-    private Double mrpPrice;
+    private int mrpPrice;
 
-    private Double sellingPrice;
+    private int sellingPrice;
 
     private Long userId;
 }
